+++ about.html
@@ -1,0 +1,155 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+    integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+  <link rel="stylesheet" href="css/styles.css">
+  <title>CSS Summit</title>
+</head>
+
+<body>
+  <aside>
+    <ul class="d-flex justify-content-end list-inline container my-1">
+      <li class="mr-2"><a href="" class="aside mr-2"><img src="images/facebook.svg" alt="A link to our facebook page"
+            id="facebook"></a></li>
+      <li class="mr-2"><a href="" class="aside"><img src="images/twitter.svg" alt="A link to our twitter page"></a></li>
+      <li><a href="" class="aside mx-2 text-decoration-none">English</a></li>
+      <li><a href="" class="aside mx-2 text-decoration-none">My Page</a></li>
+      <li><a href="" class="aside mx-2 text-decoration-none">Logout</a></li>
+    </ul>
+  </aside>
+  <nav class="navbar nav-list navbar-expand-sm navbar-light bg-light">
+    <div class="container logo font-weight-bolder">
+      <a class="navbar-brand d-flex w-auto" href="#"><span>The</span> <img src="/images/css.png" width="50" height="50"
+          alt="" class="align-self-center px-1 mb-sm-3"> <span class="align-self-end mb-2">Summit</span>
+        <span class="position-absolute mt-4">2021</span>
+      </a>
+      <button class="navbar-toggler d-lg-none" type="button" data-toggle="collapse" data-target="#collapsibleNavId"
+        aria-controls="collapsibleNavId" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="collapsibleNavId">
+        <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
+          <li class="nav-item">
+            <a class="nav-link active" href="#">About <span class="sr-only sr-only-focusable">About</span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" id="nav" href="#">Program <span class="sr-only sr-only-focusable">About</span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">Join <span class="sr-only sr-only-focusable">About</span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">Sponsor <span class="sr-only sr-only-focusable">About</span></a>
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" href="#">News<span class="sr-only sr-only-focusable">About</span></a>
+          </li>
+          <li class="nav-item nwk">
+            <a class="nav-link" href="#"><abbr class="text-decoration-none" title="Networking">NWWK</abbr><span
+                class="sr-only sr-only-focusable">
+                <abbr title="Networking">NWK</abbr></span></a>
+          </li>
+        </ul>
+      </div>
+  </nav>
+  <main class="main">
+    <article class="main-article">
+      <div class="container text-left">
+        <header class="pt-5">
+          <p class="header orange">"Hello! Innovators"</p>
+          <h1 class="header orange">
+            THE GLOBAL CSS SUMMIT 2021
+          </h1>
+          <div class="small-container border border-white bg-white my-3 p-2">
+            <small class="head-text">
+              The annual event that gathers some of the best minds from all over the world for a single purpose:
+              Cascading
+              Style Sheets!
+              We hope to gather more experts and enthusiasts this year.
+            </small>
+          </div>
+          <div class="mb-5">
+            <p>Please contact us through our email for any futher enquiries about the event.</p>
+          </div>
+          <div class="email">
+            <p>thecssummit@gmail.com</p>
+          </div>
+        </header>
+      </div>
+      <main class="container">
+        <section>
+          <div class="mt-2">
+            <h2>Have a look at the past summits</h2>
+          </div>
+          <div>
+            <p class="mb-3">Get an idea of what it is like to attend The Global CSS Summit by going through our past
+              events.</p>
+          </div>
+          <div class="past d-sm-flex">
+            <div class="summit1">
+              <div class="back-color">
+                <p class="year text-white p-3">2020</p>
+              </div>
+            </div>
+            <div class="summit2">
+              <div class="back-color">
+                <p class="year text-white p-3">2019</p>
+              </div>
+            </div>
+          </div>
+        </section>
+      </main>
+      <footer>
+        <section>
+          <div class="partner-cont mb-0 pt-3 pb-0 text-center">
+            <p class="clear">Partner</p>
+          </div>
+          <div class="dash-cont">
+            <div class="red-line mx-auto"></div>
+          </div>
+          <div class="sponsors-cont container my-4 d-flex justify-content-around">
+            <p class="clear">mozilla</p>
+            <p class="opac">Google</p>
+            <p class="clear">NAVER</p>
+            <p class="opac">facebook</p>
+            <p class="opac"><img src="images/airbnb-icon.svg" class="opac" width="15" height="15" alt=""> airbnb</p>
+          </div>
+        </section>
+        <section class="note abt-note">
+          <div class="container p-4 d-flex">
+            <div>
+              <a class="navbar-brand abt-color d-flex w-auto" href="#"><span>The</span> <img src="/images/css.png"
+                  width="50" height="50" alt="" class="align-self-center px-1 mb-sm-3"> <span
+                  class="align-self-end mb-2">Summit</span>
+                <span class="position-absolute mt-4">2021</span>
+              </a>
+            </div>
+            <div class="abt-color">
+              <p>
+                This event is fictional and no one mentioned within this site ever attended this event. The speakers
+                used are actual actual people highlited for their skills and expertise.
+              </p>
+              <p>Created and designed by <a href="https://github.com/fobadara" class="abt-color">Fobadara</a></p>
+            </div>
+          </div>
+        </section>
+      </footer>
+      <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+        crossorigin="anonymous"></script>
+      <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+        integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+        crossorigin="anonymous"></script>
+      <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+        integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+        crossorigin="anonymous"></script>
+      <script src="javascript/script.js"></script>
+</body>
+
+</html>
