+++ index.html
@@ -27,7 +27,7 @@
     <div class="container logo font-weight-bolder">
       <a class="navbar-brand d-flex w-auto" href="#"><span>The</span> <img src="/images/css.png" width="50" height="50"
           alt="" class="align-self-center px-1 mb-sm-3"> <span class="align-self-end mb-2">Summit</span>
-        <span class="position-absolute fixed-bottom ml-5 pl-5">2021</span>
+        <span class="position-absolute mt-4">2021</span>
       </a>
       <button class="navbar-toggler d-lg-none" type="button" data-toggle="collapse" data-target="#collapsibleNavId"
         aria-controls="collapsibleNavId" aria-expanded="false" aria-label="Toggle navigation">
@@ -67,7 +67,7 @@
             THE GLOBAL CSS SUMMIT 2021
           </h1>
           <div class="small-container border border-white bg-white my-3 p-2">
-            <small class="">
+            <small class="head-text">
               The annual event that gathers some of the best minds from all over the world for a single purpose:
               Cascading
               Style Sheets!
@@ -190,7 +190,7 @@
       <div class="dash-cont">
         <div class="red-line mx-auto"></div>
       </div>
-      <div class="sponsors-cont container my-4 d-flex justify-content-around">
+      <div class="container my-4 d-flex flex-wrap justify-content-around">
         <p class="clear">mozilla</p>
         <p class="opac">Google</p>
         <p class="clear">NAVER</p>
@@ -198,20 +198,33 @@
         <p class="opac"><img src="images/airbnb-icon.svg" class="opac" width="15" height="15" alt=""> airbnb</p>
       </div>
     </section>
-    <section class="note d-flex bg-white ">
-      <div>
-        <a class="navbar-brand d-flex link w-auto" href="#"><span>The</span> <img src="/images/css.png" width="50"
-            height="50" alt="" class="align-self-center px-1 mb-sm-3"> <span class="align-self-end mb-2">Summit</span>
-          <span class="position-absolute fixed-bottom ml-5 pl-5">2021</span>
-        </a>
-      </div>
-      <div>
-        <p>This event is fictional and no one mentioned within this site ever attended this event. The speakers used are
-          actual actual people highlited for their skills and expertise.</p>
-        <p>Created and designed by <a href="https://github.com/fobadara" class="link">Fobadara</a></p>
+    <section class="note  bg-white ">
+      <div class="d-flex">
+        <div>
+          <a class="container navbar-brand d-flex link w-auto" href="#"><span>The</span> <img src="/images/css.png"
+              width="50" height="50" alt="" class="align-self-center px-1 mb-sm-3"> <span
+              class="align-self-end mb-2">Summit</span>
+            <span class="position-absolute mt-4">2021</span>
+          </a>
+        </div>
+        <div>
+          <p>This event is fictional and no one mentioned within this site ever attended this event. The speakers used
+            are
+            actual actual people highlited for their skills and expertise.</p>
+          <p>Created and designed by <a href="https://github.com/fobadara" class="link">Fobadara</a></p>
+        </div>
       </div>
     </section>
   </footer>
+  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+    integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+    crossorigin="anonymous"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+    integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+    crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+    integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+    crossorigin="anonymous"></script>
   <script src="javascript/script.js"></script>
 </body>
