+++ javascript/model.js
@@ -1,0 +1,48 @@
+export default class Model {
+  constructor() {
+    this.data = [
+      {
+        bg: 'images/checker-pattern.jpeg',
+        image: 'images/mosh-hamedani.jpg',
+        name: 'Mosh Hamedani',
+        job: 'Teaches people coding through youtube and various other resources.',
+        expertise: 'A good software developer with years of experience to his belt and lots of students accross the world.',
+      },
+      {
+        bg: 'images/checker-pattern.jpeg',
+        image: 'images/wes-boss.jpg',
+        name: 'Wes Boss',
+        job: 'Teaches people coding through youtube and various other resources.',
+        expertise: 'A good software developer with years of experience to his belt and lots of students accross the world.',
+      },
+      {
+        bg: 'images/checker-pattern.jpeg',
+        image: 'images/garry-simon.jpg',
+        name: 'Garry Simon',
+        job: 'Teaches people coding through youtube and various other resources.',
+        expertise: 'A good software developer with years of experience to his belt and lots of students accross the world.',
+      },
+      {
+        bg: 'images/checker-pattern.jpeg',
+        image: 'images/dev-ed.jpg',
+        name: 'Dev-ed',
+        job: 'Teaches people coding through youtube and various other resources.',
+        expertise: 'A good software developer with years of experience to his belt and lots of students accross the world.',
+      },
+      {
+        bg: 'images/checker-pattern.jpeg',
+        image: 'images/adrian-twarog.jpg',
+        name: 'Adrian Twarog',
+        job: 'Teaches people coding through youtube and various other resources.',
+        expertise: 'A good software developer with years of experience to his belt and lots of students accross the world.',
+      },
+      {
+        bg: 'images/checker-pattern.jpeg',
+        image: 'images/kevin-powell.jpeg',
+        name: 'Kevin Powell',
+        job: 'Teaches people coding through youtube and various other resources.',
+        expertise: 'A good software developer with years of experience to his belt and lots of students accross the world.',
+      },
+    ];
+  }
+}
